live_transcription/recall_test/requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
Flask>=2.0.0  # Only needed if you're running the webhook server
